Finally finished up the UI, so we can actually start writing some classes.  However, a few things changed in the process:
I changed the Sources class to use an enum and a small array on the backend to handle all of the possible plugins.  I think this is much cleaner than how it was setup before.  Also, plugins are created/destroyed when they are enabled/disabled, as we discussed.

One thing that I'm still not happy with in the sources plugin is the fact that we pass explicit references to the Status object around to each plugin.  I think this is a pretty big break in encapsulation.  We'd be better off having another event handler, since most of this processing happens in the event thread anyway.  I know that events are kind of just a smooth way over breaking encapsulation anyway, but it just feels better.

Also, I ran Eclipse's auto-cleanup commands on all the source I changed, so the the diff's might not be pretty...